requirements.txt
langchain
langchain-groq
sentence-transformers
pytube
youtube-transcript-api
faiss-cpu
python-dotenv
fastapi
uvicorn
pydantic
requests
beautifulsoup4
langchain-community
wordcloud
matplotlib